SGP30 improvement. Added Debug SGP30 measurement

+++ pi-utils/timers.h
@@ -57,7 +57,7 @@ public:
         return nanosleep(&tm, &tm);
     }
 
-    static int milliseconds(){
+    static long milliseconds(){
         timesp tm;
         clock_gettime(CLOCK_MONOTONIC, &tm);
         return (tm.tv_sec*1000+tm.tv_nsec/1000000);
